Updating documentation, reorganizing the repo.

+++ local-install.sh
@@ -2,7 +2,12 @@
 
 set -e
 
-export PATH=$HOME/local_python/root/bin:$PATH
+if [ -z $LOCAL_PYTHON3_BIN ] ; then
+    echo "Please define LOCAL_PYTHON3_BIN environment variable."
+    exit 1
+fi
+
+export PATH=$LOCAL_PYTHON3_BIN:$PATH
 
 pip3 uninstall -y f1-2019-telemetry
