+++ make-wheel.sh
@@ -2,7 +2,12 @@
 
 set -e
 
-export PATH=$HOME/local_python/root/bin:$PATH
+if [ -z $LOCAL_PYTHON3_BIN ] ; then
+    echo "Please define LOCAL_PYTHON3_BIN environment variable."
+    exit 1
+fi
+
+export PATH=$LOCAL_PYTHON3_BIN:$PATH
 
 # Clean up
 rm -rf build dist f1_2019_telemetry.egg-info *.whl
@@ -11,7 +16,8 @@ python3 setup.py bdist_wheel
 
 cp dist/*.whl .
 
-# Clean up
+# Clean up everything except wheel file
 rm -rf build dist f1_2019_telemetry.egg-info
 
+# Show contents of wheel file
 unzip -v *.whl
